Rename eager cache to persistent cache
