+++ wcfsetup/install/files/lib/system/cache/persistent/CoreObjectCache.class.php
@@ -1,22 +1,22 @@
 <?php
 
-namespace wcf\system\cache\eager;
+namespace wcf\system\cache\persistent;
 
 use Symfony\Contracts\Cache\ItemInterface;
 use wcf\data\core\object\CoreObjectList;
 use wcf\system\SingletonFactory;
 
 /**
- * Eager cache implementation for core objects.
+ * Persistent cache implementation for core objects.
  *
  * @author Olaf Braun
  * @copyright 2001-2025 WoltLab GmbH
  * @license GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
  * @since 6.2
  *
- * @extends AbstractEagerCache<array<string, class-string<SingletonFactory>>>
+ * @extends AbstractPersistentCache<array<string, class-string<SingletonFactory>>>
  */
-final class CoreObjectCache extends AbstractEagerCache
+final class CoreObjectCache extends AbstractPersistentCache
 {
     #[\Override]
     public function __invoke(ItemInterface $item): array
