+++ wcfsetup/install/files/lib/system/language/LanguageFactory.class.php
@@ -7,8 +7,8 @@ use Negotiation\LanguageNegotiator;
 use wcf\data\language\category\LanguageCategory;
 use wcf\data\language\Language;
 use wcf\data\language\LanguageEditor;
-use wcf\system\cache\eager\data\LanguageCacheData;
-use wcf\system\cache\eager\LanguageCache;
+use wcf\system\cache\persistent\data\LanguageCacheData;
+use wcf\system\cache\persistent\LanguageCache;
 use wcf\system\SingletonFactory;
 use wcf\system\template\TemplateScriptingCompiler;
 use wcf\system\WCF;
